# api_documentation.json
{"openapi":"3.1.0","info":{"title":"Drop-off Points API","summary":"Drop-off Points API for UAchado App","description":"This API manages the drop-off points in UAchado system. It helps with the logic inside the system.","version":"1.0.0"},"paths":{"/points/v1":{"get":{"tags":["Points"],"summary":"Base","description":"Root endpoint of the dropoff points API. Mostly used for testing API connectivity.\n\nReturns:\n    _type_: A dictionary containing the response message.","operationId":"base_points_v1_get","responses":{"200":{"description":"Root endpoint of the dropoff points API.","content":{"application/json":{"schema":{"type":"object","title":"Response Base Points V1 Get"}}}}}}},"/points/v1/points":{"get":{"tags":["Points"],"summary":"Get All Points","description":"Get the list of existing points.\n\nArgs:\n    db (Session, optional): Optional database session object. If not included the system will connect to the default one. Defaults to Depends(get_db).\n\nReturns:\n    List[schemas.Point]: A list of existing points.","operationId":"get_all_points_points_v1_points_get","responses":{"200":{"description":"Get the list of existing points.","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Point"},"type":"array","title":"Response Get All Points Points V1 Points Get"}}}}}},"post":{"tags":["Points"],"summary":"Create Point","description":"Create a new point.\n\nArgs:\n    request (Request): The request object containing information about the request.\n    point (schemas.PointCreate): The payload data to create a new point.\n    db (Session, optional): Optional database session object. If not included the system will connect to the default one. Defaults to Depends(get_db).\n\nRaises:\n    HTTPException (HTTP_409_CONFLICT): Error raised if there's already a stored point with that name. \n\nReturns:\n    _type_: The created point.","operationId":"create_point_points_v1_points_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PointCreate"}}},"required":true},"responses":{"201":{"description":"Create a new point.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Point"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/points/v1/points/name/{point_name}":{"get":{"tags":["Points"],"summary":"Get Point","description":"Get a specific point by its name.\n\nArgs:\n    point_name (str): The name attribute of a specify unique point.\n    db (Session, optional): Optional database session object. If not included the system will connect to the default one. Defaults to Depends(get_db).\n\nRaises:\n    HTTPException (HTTP_204_NO_CONTENT): Error raised if there's no stored point with that name.\n\nReturns:\n    schemas.Point: The stored point.","operationId":"get_point_points_v1_points_name__point_name__get","parameters":[{"name":"point_name","in":"path","required":true,"schema":{"type":"string","title":"Point Name"}}],"responses":{"200":{"description":"Get a specific point by its name.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Point"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Points"],"summary":"Delete Point","description":"Delete a specific point by its name.\n\nArgs:\n    request (Request): The request object containing information about the request.\n    point_name (str): The name attribute of a specify unique point.\n    db (Session, optional): Optional database session object. If not included the system will connect to the default one. Defaults to Depends(get_db).\n\nRaises:\n    HTTPException (HTTP_204_NO_CONTENT): Error raised if there's no stored point with that name.\n\nReturns:\n    _type_: A dictionary containing a success message.","operationId":"delete_point_points_v1_points_name__point_name__delete","parameters":[{"name":"point_name","in":"path","required":true,"schema":{"type":"string","title":"Point Name"}}],"responses":{"200":{"description":"Delete a specific point by its name.","content":{"application/json":{"schema":{"type":"object","title":"Response Delete Point Points V1 Points Name  Point Name  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/points/v1/access":{"get":{"tags":["Points"],"summary":"Get Point Id Of Access","description":"Get the access name and drop-off point ID from the access token.\n\nArgs:\n    request (Request): request object containing information about the request.\n    db (Session, optional): Optional database session object. If not included the system will connect to the default one. Defaults to Depends(get_db).\n\nRaises:\n    HTTPException (HTTP_204_NO_CONTENT): Error raised if the tokens provided don't associate to any stored access.\n\nReturns:\n    Optional[dict]: A dictionary containing the name and the drop-off point ID of the access.","operationId":"get_point_id_of_access_points_v1_access_get","responses":{"200":{"description":"Get the access name and drop-off point ID from the access token.","content":{"application/json":{"schema":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Response Get Point Id Of Access Points V1 Access Get"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Point":{"properties":{"name":{"type":"string","title":"Name"},"location":{"type":"string","title":"Location"},"coordinates":{"type":"string","title":"Coordinates"},"image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","location","coordinates","image","id"],"title":"Point","description":"Represents a point in a 2D coordinate system.\n\nAttributes:\n    id (int): The unique identifier for the point."},"PointCreate":{"properties":{"name":{"type":"string","title":"Name"},"location":{"type":"string","title":"Location"},"coordinates":{"type":"string","title":"Coordinates"},"image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image"}},"type":"object","required":["name","location","coordinates","image"],"title":"PointCreate","description":"A class representing a point created by the user.\n\nThis class inherits from the PointBase class.\n\n.. note::\n    The PointCreate class is used to represent points that are created by the user.\n\n.. seealso::\n    :py:class:`PointBase`"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}